<fix>[compute]: introduce HostCandidateProducer
HostCandidateProducer is separated from AbstractHostAllocatiorFlow
and is specifically used to preliminarily select a list of hosts
that meet the requirements.

After splitting, AbstractHostAllocatiorFlow will no longer hold
pagination information and will not need to know if it is
the first element in the flows list.

The steps of original host allocator are as follows:

```
AbstractHostAllocatorFlow (producer + consumer)
-> AbstractHostSortorFlow
```

The steps of updated host allocator are as follows:

```
HostCandidateProducer (producer)
-> AbstractHostAllocatorFlow (consumer)
-> AbstractHostSortorFlow
```

Related: ZSV-7739

Change-Id: I726b677372756a786f68646562716f757a626561